LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/require-corp.https-expected.txt


FAIL "require-corp" top-level: navigating a frame to "none" should fail assert_equals: expected null but got Document node with 1 child
FAIL "require-corp" top-level: navigating a frame from "require-corp" to "none" should fail Can't find variable: BroadcastChannel
FAIL "require-corp" top-level: creating a noopener "none" popup should succeed Can't find variable: BroadcastChannel
FAIL "require-corp" top-level: creating a "none" popup should succeed. Can't find variable: BroadcastChannel
FAIL "require-corp" top-level (as popup): navigating to "none" should succeed Can't find variable: BroadcastChannel
FAIL "require-corp" top-level (as noopener popup): navigating to "none" should succeed Can't find variable: BroadcastChannel
FAIL "require-corp" top-level (as popup with opener set to null): navigating to "none" should succeed Can't find variable: BroadcastChannel
PASS "require-corp" top-level: fetch() to CORP: cross-origin response should succeed 
FAIL "require-corp" top-level: fetch() to response without CORP should fail assert_unreached: Should have rejected: undefined Reached unreachable code
